Move screens to corresponding packages
