+++ android/app/src/main/java/ru/spbstu/amd/learnbraille/screens/exit/ExitFragment.kt
@@ -1,10 +1,11 @@
-package ru.spbstu.amd.learnbraille
+package ru.spbstu.amd.learnbraille.screens.exit
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
+import ru.spbstu.amd.learnbraille.R
 import ru.spbstu.amd.learnbraille.databinding.FragmentExitBinding
 import kotlin.system.exitProcess
 
@@ -13,7 +14,8 @@ class ExitFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
     ) = DataBindingUtil.inflate<FragmentExitBinding>(
-        inflater, R.layout.fragment_exit, container, false
+        inflater,
+        R.layout.fragment_exit, container, false
     ).apply {
         exitButton.setOnClickListener {
             exitProcess(0)
