+++ android/app/src/main/java/ru/spbstu/amd/learnbraille/screens/greeting/GreetingFragment.kt
@@ -1,4 +1,4 @@
-package ru.spbstu.amd.learnbraille
+package ru.spbstu.amd.learnbraille.screens.greeting
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,6 +6,7 @@ import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.navigation.Navigation
+import ru.spbstu.amd.learnbraille.R
 import ru.spbstu.amd.learnbraille.databinding.FragmentGreetingBinding
 
 class GreetingFragment : Fragment() {
@@ -13,7 +14,8 @@ class GreetingFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
     ) = DataBindingUtil.inflate<FragmentGreetingBinding>(
-        inflater, R.layout.fragment_greeting, container, false
+        inflater,
+        R.layout.fragment_greeting, container, false
     ).apply {
         startButton.setOnClickListener(
             Navigation.createNavigateOnClickListener(R.id.action_greetingFragment_to_menuFragment)
