+++ android/app/src/main/java/ru/spbstu/amd/learnbraille/screens/menu/MenuFragment.kt
@@ -1,4 +1,4 @@
-package ru.spbstu.amd.learnbraille
+package ru.spbstu.amd.learnbraille.screens.menu
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,6 +6,7 @@ import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.navigation.Navigation
+import ru.spbstu.amd.learnbraille.R
 import ru.spbstu.amd.learnbraille.databinding.FragmentMenuBinding
 
 class MenuFragment : Fragment() {
@@ -13,7 +14,8 @@ class MenuFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
     ) = DataBindingUtil.inflate<FragmentMenuBinding>(
-        inflater, R.layout.fragment_menu, container, false
+        inflater,
+        R.layout.fragment_menu, container, false
     ).apply {
         practiceButton.setOnClickListener(
             Navigation.createNavigateOnClickListener(R.id.action_menuFragment_to_practiceFragment)
