+++ android/app/src/main/res/layout/fragment_menu.xml
@@ -6,7 +6,7 @@
         android:id="@+id/constraintMenuLayout"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MenuFragment">
+        tools:context=".screens.menu.MenuFragment">
 
 
         <Button
