+++ android/app/src/main/res/layout/fragment_practice.xml
@@ -7,7 +7,7 @@
         android:id="@+id/constraintLayout"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".PracticeFragment">
+        tools:context=".screens.practice.PracticeFragment">
 
         <TextView
                 android:id="@+id/letter"
